requirements.txt
fastapi~=0.112.2
pydantic-settings~=2.6.1
uvicorn[standard]
sqlalchemy~=2.0.37
databases[aiosqlite]~=0.9.0
python-dotenv
httpx~=0.28.1
rich
asgi-correlation-id~=4.3.4
python-json-logger
logtail-python
python-jose~=3.3.0
python-multipart
passlib[bcrypt]~=1.7.4
vine==5.1.0
celery>=5.4.0
redis
aiofiles~=22.1.0
b2sdk~=2.8.0
pytest~=8.3.4
pydantic~=2.10.3